refactor: put process helpers outside
